- the `pom.xml` now looks like any other Maven build generated from the [Spring Initializr](http://start.spring.io) - moved the `Backend` and `Frontend` to separate packages so that the auto-configuration of one module doesn't interact with the other - preferred constructor injection for the `RestTemplate`. The `Resttemplate` now uses `@Lazy` so that the Zipkin auto-configuration can interact with the produced `RestTemplate` and so that we can inject the required bean in our `@RestController` - Spring Boot makes it trivial to standup a Zipkin-server so this example, breaking with the others, _also_ includes a Zipkin server implementation.
