+++ src/main/java/sleuth/webmvc/backend/Backend.java
@@ -1,10 +1,11 @@
-package sleuth.webmvc;
+package sleuth.webmvc.backend;
 
-import java.util.Date;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Date;
 
 @SpringBootApplication
 @RestController
