+++ src/main/java/sleuth/webmvc/server/ZipkinServer.java
@@ -1,0 +1,14 @@
+package sleuth.webmvc.server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import zipkin.server.EnableZipkinServer;
+
+@SpringBootApplication
+@EnableZipkinServer
+public class ZipkinServer {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ZipkinServer.class, "--server.port=9411");
+	}
+}
